+++ include/rapidjson/writer.h
@@ -350,7 +350,7 @@ template<>
 inline bool Writer<StringBuffer>::WriteInt(int i) {
     char *buffer = os_->Push(11);
     const char* end = internal::i32toa(i, buffer);
-    os_->Pop(11 - (end - buffer));
+    os_->Pop(static_cast<size_t>(11 - (end - buffer)));
     return true;
 }
 
@@ -358,7 +358,7 @@ template<>
 inline bool Writer<StringBuffer>::WriteUint(unsigned u) {
     char *buffer = os_->Push(10);
     const char* end = internal::u32toa(u, buffer);
-    os_->Pop(10 - (end - buffer));
+    os_->Pop(static_cast<size_t>(10 - (end - buffer)));
     return true;
 }
 
@@ -366,7 +366,7 @@ template<>
 inline bool Writer<StringBuffer>::WriteInt64(int64_t i64) {
     char *buffer = os_->Push(21);
     const char* end = internal::i64toa(i64, buffer);
-    os_->Pop(21 - (end - buffer));
+    os_->Pop(static_cast<size_t>(21 - (end - buffer)));
     return true;
 }
 
@@ -374,7 +374,7 @@ template<>
 inline bool Writer<StringBuffer>::WriteUint64(uint64_t u) {
     char *buffer = os_->Push(20);
     const char* end = internal::u64toa(u, buffer);
-    os_->Pop(20 - (end - buffer));
+    os_->Pop(static_cast<size_t>(20 - (end - buffer)));
     return true;
 }
 
@@ -382,7 +382,7 @@ template<>
 inline bool Writer<StringBuffer>::WriteDouble(double d) {
     char *buffer = os_->Push(25);
     char* end = internal::dtoa(d, buffer);
-    os_->Pop(25 - (end - buffer));
+    os_->Pop(static_cast<size_t>(25 - (end - buffer)));
     return true;
 }
